image_caption/requirements.txt
transformers
gradio
opencv-python
pillow
tensorflow
torch
tf-keras
